Refactor the goroutine code to internal helpers
