+++ data/binding/queue.go
@@ -1,30 +1,12 @@
 package binding
 
 import (
-	"runtime"
-
 	"fyne.io/fyne/v2"
+	"fyne.io/fyne/v2/internal/async"
 )
 
-var mainGoroutineID uint64
-
-func init() {
-	runtime.LockOSThread()
-	mainGoroutineID = goroutineID()
-}
-
-func goroutineID() (id uint64) {
-	var buf [30]byte
-	runtime.Stack(buf[:], false)
-	for i := 10; buf[i] != ' '; i++ {
-		id = id*10 + uint64(buf[i]&15)
-	}
-
-	return id
-}
-
 func queueItem(f func()) {
-	if goroutineID() == mainGoroutineID {
+	if async.IsMainGoroutine() {
 		f()
 		return
 	}
