+++ internal/async/goroutine.go
@@ -1,0 +1,11 @@
+package async
+
+var mainGoroutineID uint64
+
+func init() {
+	mainGoroutineID = goroutineID()
+}
+
+func IsMainGoroutine() bool {
+	return goroutineID() == mainGoroutineID
+}
