+++ internal/async/goroutine_desktop.go
@@ -1,0 +1,15 @@
+//go:build !wasm
+
+package async
+
+import "runtime"
+
+func goroutineID() (id uint64) {
+	var buf [30]byte
+	runtime.Stack(buf[:], false)
+	for i := 10; buf[i] != ' '; i++ {
+		id = id*10 + uint64(buf[i]&15)
+	}
+
+	return id
+}
