+++ internal/driver/glfw/loop.go
@@ -27,7 +27,6 @@ var initOnce = &sync.Once{}
 // Arrange that main.main runs on main thread.
 func init() {
 	runtime.LockOSThread()
-	mainGoroutineID = goroutineID()
 }
 
 // force a function f to run on the main thread
